refactor: Router
Refactor Router component
