+++ .github/workflows/ci.yml
@@ -20,14 +20,6 @@ jobs:
           cache: 'npm'
           node-version-file: '.nvmrc'
 
-      - name: 💾 Cache node modules
-        uses: actions/cache@v3
-        with:
-          path: ~/.npm
-          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
-          restore-keys: |
-            ${{ runner.os }}-node-
-
       - name: 🔋 Install
         run: npm ci
