--- src/components/Card/style.module.css
@@ -1,3 +1,0 @@
-.card {
-  position: relative;
-}
