--- src/components/Card/style.module.css.d.ts
@@ -1,4 +1,0 @@
-declare const styles: {
-  readonly card: string;
-};
-export = styles;
