--- src/pages/Index/index.tsx
@@ -1,15 +1,0 @@
-import { Helmet } from 'react-helmet';
-import { Navigate } from 'react-router-dom';
-
-function IndexPage(): JSX.Element {
-  return (
-    <>
-      <Helmet>
-        <title>Weather App</title>
-      </Helmet>
-      <Navigate to="/cities" />
-    </>
-  );
-}
-
-export default IndexPage;
